U.Core.ABT

Documentation

data ABT f v r

Constructors

Var v

Cycle r

Abs v r

Tm (f r)

data Term f v a

At each level in the tree, we store the set of free variables and a value of type a. Variables are of type v.

Constructors

Term

Fields

freeVars :: Set v

annotation :: a

out :: ABT f v (Term f v a)

data Term' f v a x

Constructors

Term'

Fields

freeVars' :: Set v

annotation' :: a

out' :: ABT f v x

amap :: Functor f => (a -> a') -> Term f v a -> Term f v a'

vmap :: (Functor f, Foldable f, Ord v') => (v -> v') -> Term f v a -> Term f v' a

vmapM :: (Applicative m, Traversable f, Foldable f, Ord v2) => (v -> m v2) -> Term f v a -> m (Term f v2 a)

transform :: (Ord v, Foldable g, Functor g) => (forall a. f a -> g a) -> Term f v a -> Term g v a

transformM :: (Ord v, Monad m, Traversable g) => (forall a. f a -> m (g a)) -> Term f v a -> m (Term g v a)

abs :: Ord v => a -> v -> Term f v a -> Term f v a

var :: a -> v -> Term f v a

cycle :: a -> Term f v a -> Term f v a

tm :: (Foldable f, Ord v) => a -> f (Term f v a) -> Term f v a

Traversals

visit :: (Traversable f, Applicative g, Ord v) => (Term f v a -> Maybe (g (Term f v a))) -> Term f v a -> g (Term f v a)

`visit f t` applies an effectful function to each subtree of t and sequences the results. When f returns Nothing, visit descends into the children of the current subtree. When f returns `Just t2`, visit replaces the current subtree with t2. Thus: `visit (const Nothing) t == pure t` and `visit (const (Just (pure t2))) t == pure t2`

visit' :: (Traversable f, Monad g, Ord v) => (f (Term f v a) -> g (f (Term f v a))) -> Term f v a -> g (Term f v a)

Apply an effectful function to an ABT tree top down, sequencing the results.

visit_ :: (Traversable f, Applicative g, Ord v) => (f (Term f v a) -> g ()) -> Term f v a -> g (Term f v a)

Apply an effectful function to an ABT tree top down, sequencing the results.

visitPure :: (Traversable f, Ord v) => (Term f v a -> Maybe (Term f v a)) -> Term f v a -> Term f v a

visit specialized to the Identity effect.

foreachSubterm :: (Traversable f, Applicative g) => (Term f v a -> g b) -> Term f v a -> g [b]

subterms :: (Ord v, Traversable f) => Term f v a -> [Term f v a]

Patterns

pattern Var' :: v -> Term f v a

pattern Cycle' :: [v] -> Term f v a -> Term f v a

pattern AbsN' :: [v] -> Term f v a -> Term f v a

pattern Tm' :: f (Term f v a) -> Term f v a

unabs :: Term f v a -> ([v], Term f v a)

freshInBoth :: Var v => Term f v a -> Term f v a -> v -> v

Produce a variable which is free in both terms

substsInheritAnnotation :: (Foldable f, Functor f, Var v) => [(v, Term f v b)] -> Term f v a -> Term f v a

substInheritAnnotation :: (Foldable f, Functor f, Var v) => v -> Term f v b -> Term f v a -> Term f v a

subst' :: (Foldable f, Functor f, Var v) => (a -> Term f v a) -> v -> Set v -> Term f v a -> Term f v a

rename :: (Foldable f, Functor f, Var v) => v -> v -> Term f v a -> Term f v a

renames old to new in the given term, ignoring subtrees that bind old

allVars :: Foldable f => Term f v a -> [v]

freshenS :: (Var v, MonadState (Set v) m) => v -> m v

Freshens the given variable wrt. the set of used variables tracked by state. Adds the result to the set of used variables.

freshenS' :: (Var v, MonadState s m) => Lens' s (Set v) -> v -> m v

A more general version of freshenS that uses a lens to focus on used variables inside state.