U.Core.ABT.Var

Documentation

class Ord v => Var v where

A class for avoiding accidental variable capture

`Set.notMember (freshIn vs v) vs`: freshIn returns a variable not used in the Set

Methods

freshIn :: Set v -> v -> v

substsInheritAnnotation :: (Foldable f, Functor f, Var v) => [(v, Term f v b)] -> Term f v a -> Term f v a

substInheritAnnotation :: (Foldable f, Functor f, Var v) => v -> Term f v b -> Term f v a -> Term f v a

subst' :: (Foldable f, Functor f, Var v) => (a -> Term f v a) -> v -> Set v -> Term f v a -> Term f v a

rename :: (Foldable f, Functor f, Var v) => v -> v -> Term f v a -> Term f v a

renames old to new in the given term, ignoring subtrees that bind old