Unison.Core.Project

Description

Project-related types.

A larger API, including orphan instances for parsing from Text, is available in Unison.Project. Here, we just define the types, which are shared among the low-level database layer (which assumes without verifying that project names and such are syntactically valid) and the higher-level project manipulation exposed by UCM.

Documentation

newtype ProjectName

The name of a project.

Constructors

UnsafeProjectName Text

newtype ProjectBranchName

The name of a branch of a project.

Constructors

UnsafeProjectBranchName Text

data ProjectAndBranch a b

A generic data structure that contains information about a project and a branch in that project.

Constructors

ProjectAndBranch

Fields

project :: a

branch :: b