U.Codebase.HashTags

Documentation

newtype ComponentHash

Represents a hash of a type or term component

Constructors

ComponentHash

Fields

unComponentHash :: Hash

newtype BranchHash

Constructors

BranchHash

Fields

unBranchHash :: Hash

newtype CausalHash

Represents a hash of a causal containing values of the provided type.

Constructors

CausalHash

Fields

unCausalHash :: Hash

newtype PatchHash

Constructors

PatchHash

Fields

unPatchHash :: Hash